refactor url.ast.data.predicate to model.predicate to avoid circular deps [ci skip]
