+++ ermrest/model/__init__.py
@@ -30,6 +30,7 @@ from .table import Table
 from .schema import Model, Schema
 from .introspect import introspect, current_model_version
 from . import name
+from . import predicate
 
-__all__ = ["introspect", "current_model_version", "Model", "Schema", "Table", "Column", "Type", "name"]
+__all__ = ["introspect", "current_model_version", "Model", "Schema", "Table", "Column", "Type", "name", "predicate"]
