+++ ermrest/model/makefile-vars
@@ -6,6 +6,7 @@ ERMREST_MODEL_PYTHON_FILES= \
 	key.py \
 	misc.py \
 	name.py \
+	predicate.py \
 	schema.py \
 	table.py \
 	type.py
