+++ ermrest/model/name.py
@@ -1,6 +1,6 @@
 
 # 
-# Copyright 2013-2016 University of Southern California
+# Copyright 2013-2017 University of Southern California
 # 
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -28,6 +28,7 @@ import web
 
 from ..util import sql_identifier, sql_literal
 from .. import exception
+from .predicate import Value
 
 def _keyref_join_str(self, refop, lname, rname):
     if refop == '=@':
