+++ ermrest/url/ast/data/makefile-vars
@@ -1,8 +1,7 @@
 
 ERMREST_URL_AST_DATA_PYTHON_FILES= \
 	__init__.py \
-	path.py \
-	predicate.py
+	path.py
 
 ERMREST_URL_AST_DATA_PYTHON_FILES_INSTALL=$(ERMREST_URL_AST_DATA_PYTHON_FILES:%=$(PYLIBDIR)/ermrest/url/ast/data/%)
 
@@ -11,6 +10,7 @@ INSTALL_FILES += $(ERMREST_URL_AST_DATA_PYTHON_FILES_INSTALL)
 INSTALL_DIRS += $(PYLIBDIR)/ermrest/url/ast/data
 
 CLEAN_FILES += \
+	$(PYLIBDIR)/ermrest/url/ast/data/predicate.py \
 	$(ERMREST_URL_AST_DATA_PYTHON_FILES:%=ermrest/url/ast/data/%c) \
 	$(ERMREST_URL_AST_DATA_PYTHON_FILES:%=ermrest/url/ast/data/%o)
